+++ src/main/java/com/damoa/domain/community/repository/CommunityMemberRepository.java
@@ -1,7 +1,7 @@
 package com.damoa.domain.community.repository;
 
 import com.damoa.domain.community.entity.CommunityMember;
-import com.damoa.domain.member.Member;
+import com.damoa.domain.member.entity.Member;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.EntityGraph;
