+++ src/main/java/com/damoa/domain/member/entity/Member.java
@@ -1,8 +1,8 @@
-package com.damoa.domain.member;
+package com.damoa.domain.member.entity;
 
 import com.damoa.domain.BaseTimeEntity;
 import com.damoa.web.dto.common.UploadFile;
-import com.damoa.web.dto.member.UpdateMemberDto;
+import com.damoa.domain.member.dto.UpdateMemberDto;
 import lombok.*;
 
 import javax.persistence.*;
