+++ src/main/java/com/damoa/domain/member/entity/MemberRole.java
@@ -1,4 +1,4 @@
-package com.damoa.domain.member;
+package com.damoa.domain.member.entity;
 
 
 public enum MemberRole {
