+++ src/main/java/com/damoa/domain/member/repository/MemberRepository.java
@@ -1,6 +1,6 @@
-package com.damoa.repository;
+package com.damoa.domain.member.repository;
 
-import com.damoa.domain.member.Member;
+import com.damoa.domain.member.entity.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
