+++ src/main/java/com/damoa/domain/post/Comment.java
@@ -1,7 +1,7 @@
 package com.damoa.domain.post;
 
 import com.damoa.domain.BaseTimeEntity;
-import com.damoa.domain.member.Member;
+import com.damoa.domain.member.entity.Member;
 import lombok.*;
 
 import javax.persistence.*;
