+++ src/main/java/com/damoa/service/PostServiceImpl.java
@@ -2,7 +2,7 @@ package com.damoa.service;
 
 import com.damoa.domain.community.entity.Community;
 import com.damoa.domain.community.entity.CommunityMember;
-import com.damoa.domain.member.Member;
+import com.damoa.domain.member.entity.Member;
 import com.damoa.domain.post.Comment;
 import com.damoa.domain.post.Image;
 import com.damoa.domain.post.Post;
