+++ src/main/java/com/damoa/web/dto/post/CommentDto.java
@@ -2,7 +2,7 @@ package com.damoa.web.dto.post;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import com.fasterxml.jackson.annotation.JsonInclude;
-import com.damoa.domain.member.Member;
+import com.damoa.domain.member.entity.Member;
 import com.damoa.domain.post.Comment;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
