+++ src/main/java/com/damoa/web/dto/post/CreateCommentDto.java
@@ -1,6 +1,6 @@
 package com.damoa.web.dto.post;
 
-import com.damoa.domain.member.Member;
+import com.damoa.domain.member.entity.Member;
 import com.damoa.domain.post.Comment;
 import com.damoa.domain.post.Post;
 import lombok.*;
