+++ src/main/java/com/damoa/web/dto/post/CreatePostDto.java
@@ -1,7 +1,7 @@
 package com.damoa.web.dto.post;
 
 import com.damoa.domain.community.entity.Community;
-import com.damoa.domain.member.Member;
+import com.damoa.domain.member.entity.Member;
 import com.damoa.domain.post.Post;
 import lombok.*;
 import org.hibernate.validator.constraints.Length;
