+++ src/main/java/com/damoa/web/dto/post/PostDto.java
@@ -2,7 +2,7 @@ package com.damoa.web.dto.post;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import com.damoa.domain.community.entity.Community;
-import com.damoa.domain.member.Member;
+import com.damoa.domain.member.entity.Member;
 import com.damoa.domain.post.Post;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
