+++ src/test/java/com/damoa/service/CommentServiceTest.java
@@ -4,7 +4,7 @@ import com.damoa.domain.community.entity.Category;
 import com.damoa.domain.community.entity.Community;
 import com.damoa.domain.community.entity.CommunityMember;
 import com.damoa.domain.community.entity.MemberGrade;
-import com.damoa.domain.member.Member;
+import com.damoa.domain.member.entity.Member;
 import com.damoa.domain.post.Comment;
 import com.damoa.domain.post.Post;
 import com.damoa.repository.CommentRepository;
