+++ src/test/java/com/damoa/service/CommunityServiceTest.java
@@ -6,7 +6,7 @@ import com.damoa.domain.community.entity.CommunityMember;
 import com.damoa.domain.community.entity.MemberGrade;
 import com.damoa.domain.community.service.CategoryService;
 import com.damoa.domain.community.service.CommunityServiceImpl;
-import com.damoa.domain.member.Member;
+import com.damoa.domain.member.entity.Member;
 import com.damoa.domain.community.repository.CommunityMemberRepository;
 import com.damoa.domain.community.repository.CommunityRepository;
 import com.damoa.web.dto.common.UploadFile;
