+++ src/test/java/com/damoa/web/controller/CommentControllerTest.java
@@ -1,10 +1,10 @@
 package com.damoa.web.controller;
 
-import com.damoa.domain.member.Member;
+import com.damoa.domain.member.entity.Member;
 import com.damoa.web.common.BaseControllerTest;
 import com.damoa.domain.community.dto.CreateCommunityDto;
-import com.damoa.web.dto.member.JwtTokenDto;
-import com.damoa.web.dto.member.LoginMemberDto;
+import com.damoa.domain.member.dto.JwtTokenDto;
+import com.damoa.domain.member.dto.LoginMemberDto;
 import com.damoa.web.dto.post.CreateCommentDto;
 import com.damoa.web.dto.post.CreatePostDto;
 import com.damoa.web.dto.post.PostDto;
